Trim output for empty strings.
Due to exiv2's output not having delimiters, there is no way we can figure out
how long an empty string is (some fields in Nikon makernotes are 12 empty
spaces or 15 empty spaces). Therefor we trim the parsed value from Taglib# and
hope it equals to the empty string. This is one of those cases where generated
unit tests will not be 100% correct, yet still decently sufficient. Note that
trimming only happens if the expected output is empty.  Otherwise, values are
compared as-is.