Implement visual validation of image data.
This means that not only the metadata is checked pre/post write, but also the
image data itself. This is done by reading the image using GdkPixbuf (RAW
support still TODO) and then rendering it to a PNG without metadata.
This byte dump is then compared for the original and the modified file.

This makes the tests a bit slower. I don't care.