Fix build in Monodevelop.
This was done against a Monodevelop version from trunk, but it should work with
any reasonably new version. Unit tests are synchronized with the Makefile, the
main assembly uses a special .sources file which does not allow that trick.
Might deprecate that one at some point.

The tests makefile was modified to allow building in Monodevelop.