Introduce support for GIF files
This commit introduces support for GIF files. A GIF file is parsed
completely and XMP metadata and GIF comments are extracted. Additionally,
the size of the image is extracted. Since GIF files can be used as
an animation, it would be reasonable to determine some Properties
which are related to Audio/Video files. This is currently not covered.

https://bugzilla.gnome.org/show_bug.cgi?id=619432