Create class ImageBlockFile for image files organized with data blocks
Some image file formats are organized as data blocks which can be reordered.
Those block are called 'segments' for Jpeg/Jfif files, or 'blocks' for Gif
files. Some of the blocks are metadata and handled by TagLib#. When metadata
is saved back to file, the following must be typically done:
(1) some of the blocks have to be deleted
(2) other blocks are overwritten with the new metadata

The new class ImageBlockFile allows to mark some blocks in the file as
metadata. When metadata is written back to file, the blocks which can be reused
are overwritten, the others are deleted. The File-classes for Jpeg and
Gif can now share this functionality.
Currently, ImageBlockFile only supports saving one data block (which
may contain multiple metadata blocks) and removing all other marked
blocks in file. That is fine for the currently implemented file formats, but
may be changed in future.

https://bugzilla.gnome.org/show_bug.cgi?id=619921