tests: retrieve git submodule before running tests
Some tests may fail with FileNotFoundExceptions if the
git submodule "rawsamples" is not retrieved first. This
commits makes the "test" target clever enough to retrieve
the submodule, in a similar way it is done in Banshee
build workflow (in the latter, autogen.sh would check out
the submodule, but in this case, we retrieve it at build
time because it depends on the --enable-tests flag).